pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macos-llm-file-cleanup"
version = "26.01"
description = "Local-first macOS organizer that uses metadata-aware plugins and an optional local LLM."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "macos-llm-file-cleanup" }]
dependencies = []

[project.optional-dependencies]
pdf = ["pypdf>=4.0.0"]
docx = ["python-docx>=1.1.0"]
odt = ["odfpy>=1.4.1"]
dev = ["pytest>=7.0.0"]
image = ["pillow>=10.0.0", "pillow-heif>=0.18.0"]

[project.scripts]
macos-llm-file-cleanup = "macos_llm_file_cleanup.cli:main"
